+++ .gitignore
@@ -39,6 +39,8 @@ hs_err_pid*
 .idea/**/dictionaries
 .idea/**/shelf
 
+.idea/misc.xml
+
 # Generated files
 .idea/**/contentModel.xml
